// apps/admin/src/shared/api/generated/logs/schemas/memberLoginsParams.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * JA Korea Homepage Admin API — Logs subset
 * Filtered for Homepage Admin logs Orval codegen.
 * OpenAPI spec version: v1
 */

export type MemberLoginsParams = {
audience: string;
name?: string;
loginId?: string;
from?: string;
to?: string;
page?: number;
size?: number;
};